moved readhar to addons and adjusted tests
