+++ mitmproxy/addons/readhar.py
