+++ test/mitmproxy/addons/test_readhar.py
@@ -2,13 +2,14 @@ import json
 from pathlib import Path
 
 import pytest
-from readhar import ReadHar
+from mitmproxy.addons.readhar import ReadHar
 
 from mitmproxy import exceptions
 from mitmproxy import types
+
 from mitmproxy.tools.web.app import flow_to_json
 
-here = Path(__file__).parent.absolute()
+here = Path(__file__).parent.parent.absolute() / "data"
 
 
 def hardcode_variable_fields_for_tests(flow: dict) -> None:
@@ -20,15 +21,15 @@ def hardcode_variable_fields_for_tests(flow: dict) -> None:
 
 def file_to_flows(path_name: Path) -> list[dict]:
     r = ReadHar()
-    with open(path_name, "rb") as f:
-        file_json = json.load(f)["log"]["entries"]
-        flows = []
 
-        for entry in file_json:
-            expected = r.request_to_flow(entry)
-            flow_json = flow_to_json(expected)
-            hardcode_variable_fields_for_tests(flow_json)
-            flows.append(flow_json)
+    file_json = json.loads(path_name.read_bytes())["log"]["entries"]
+    flows = []
+
+    for entry in file_json:
+        expected = r.request_to_flow(entry)
+        flow_json = flow_to_json(expected)
+        hardcode_variable_fields_for_tests(flow_json)
+        flows.append(flow_json)
 
     return flows
 
@@ -36,12 +37,14 @@ def file_to_flows(path_name: Path) -> list[dict]:
 def test_corrupt():
     r = ReadHar()
 
-    pytest.raises(
-        exceptions.CommandError, r.read_har, types.Path("./corrupted/brokenfile.har")
+    with pytest.raises(exceptions.CommandError):
+        r.read_har(types.Path(here / "corrupted_har/brokenfile.har"))
+
+    file_json = json.loads(
+        Path(here / "corrupted_har/broken_headers.json").read_bytes()
     )
-    with open("./corrupted/broken_headers.json") as f:
-        file_json = json.load(f)
-        pytest.raises(exceptions.OptionsError, r.fix_headers, file_json["headers"])
+    with pytest.raises(exceptions.OptionsError):
+        r.fix_headers(file_json["headers"])
 
 
 @pytest.mark.parametrize(
@@ -57,6 +60,14 @@ def test_har_to_flow(har_file: Path):
         actual = json.loads(json.dumps(actual))
 
         assert actual == expected
+
+
+@pytest.mark.parametrize(
+    "har_file", [pytest.param(x, id=x.stem) for x in here.glob("har_files/*.har")]
+)
+async def test_read_har(har_file):
+    rh = ReadHar()
+    assert rh.read_har(har_file) is None
 
 
 if __name__ == "__main__":
