+++ test/mitmproxy/data/corrupted_har/broken_headers.json
