+++ test/mitmproxy/data/har_files/safari.json
